Implement the bubbling behavior of the input application phase |
This behavior comes from the recursive nature of walking
tree structures. If an entity is one of the leafs is moved
out of the bounds of the leaf during the input application
phase if is out of scope for the leaf. It must be passed to the
leaves parent which is done but returning it from the method
as the leaf's parent is the method's caller.