[filu/actor] Add actor selection/creation stream
I used a different pattern here then in the auth stream. This
pattern is better I think because it is more clear how everything
is chained together, especially from the users perspective. A negative
is that it requires each implementation of Consumer/Sink to implement
it's own concurrency, but this allows for more clarity when
a stream step starts doing more complex logic. The tests sure
look prettier then the auth package.