[filu/net/client] Remove buffering of response channels
This pattern will break select statements that are waiting
on all potential response channels because after writing the
response value all the response channels were being closed.
A Select statement waiting on all the channels now can select
any of them and proceed, though the chance of it proceeding
and reading a zero value are high.